expanded _Kib _Mib sizes

+++ src/common.hpp
@@ -8,14 +8,14 @@ namespace gbx {
 template<class T>
 using owner = T;
 
-constexpr int32_t operator"" _Kib(unsigned long long kibs)
+constexpr size_t operator"" _Kib(unsigned long long kibs)
 {
-	return static_cast<int32_t>(kibs * 1024);
+	return static_cast<size_t>(kibs * 1024);
 }
 
-constexpr int32_t operator"" _Mib(unsigned long long mibs)
+constexpr size_t operator"" _Mib(unsigned long long mibs)
 {
-	return static_cast<int32_t>(mibs * 1024 * 1024);
+	return static_cast<size_t>(mibs * 1024 * 1024);
 }
 
 template<class T>
